# Conflicts:
#	src/octoprint/plugins/announcements/static/js/announcements.js
#	src/octoprint/plugins/pluginmanager/static/js/pluginmanager.js
#	src/octoprint/plugins/softwareupdate/static/js/softwareupdate.js
#	src/octoprint/server/__init__.py
#	src/octoprint/static/css/octoprint.css
#	src/octoprint/static/js/app/viewmodels/connection.js
#	src/octoprint/static/js/app/viewmodels/control.js
#	src/octoprint/static/js/app/viewmodels/files.js
#	src/octoprint/static/js/app/viewmodels/gcode.js
#	src/octoprint/static/js/app/viewmodels/log.js
#	src/octoprint/static/js/app/viewmodels/loginstate.js
#	src/octoprint/static/js/app/viewmodels/navigation.js
#	src/octoprint/static/js/app/viewmodels/printerprofiles.js
#	src/octoprint/static/js/app/viewmodels/printerstate.js
#	src/octoprint/static/js/app/viewmodels/settings.js
#	src/octoprint/static/js/app/viewmodels/temperature.js
#	src/octoprint/static/js/app/viewmodels/terminal.js
#	src/octoprint/static/js/app/viewmodels/timelapse.js
#	src/octoprint/static/js/app/viewmodels/users.js

+++ src/octoprint/plugins/softwareupdate/updaters/pip.py
@@ -15,6 +15,8 @@ from .. import exceptions
 logger = logging.getLogger("octoprint.plugins.softwareupdate.updaters.pip")
 console_logger = logging.getLogger("octoprint.plugins.softwareupdate.updaters.pip.console")
 
+_ALREADY_INSTALLED = "Requirement already satisfied (use --upgrade to upgrade)"
+
 _pip_callers = dict()
 _pip_version_dependency_links = pkg_resources.parse_version("1.5")
 
@@ -35,7 +37,7 @@ def _get_pip_caller(command=None):
 
 	return _pip_callers[key]
 
-def perform_update(target, check, target_version, log_cb=None, online=True):
+def perform_update(target, check, target_version, log_cb=None, online=True, force=False):
 	pip_command = None
 	if "pip_command" in check:
 		pip_command = check["pip_command"]
@@ -66,10 +68,10 @@ def perform_update(target, check, target_version, log_cb=None, online=True):
 		pip_caller.on_log_stdout = _log_stdout
 		pip_caller.on_log_stderr = _log_stderr
 
-	install_arg = check["pip"].format(target_version=target_version)
+	install_arg = check["pip"].format(target_version=target_version, target=target_version)
 
 	logger.debug(u"Target: %s, executing pip install %s" % (target, install_arg))
-	pip_args = ["install", check["pip"].format(target_version=target_version, target=target_version)]
+	pip_args = ["install", install_arg]
 
 	if "dependency_links" in check and check["dependency_links"]:
 		pip_args += ["--process-dependency-links"]
@@ -77,12 +79,17 @@ def perform_update(target, check, target_version, log_cb=None, online=True):
 	returncode, stdout, stderr = pip_caller.execute(*pip_args)
 	if returncode != 0:
 		raise exceptions.UpdateError("Error while executing pip install", (stdout, stderr))
+	
+	if not force and any(map(lambda x: x.strip().startswith(_ALREADY_INSTALLED) and (install_arg in x or install_arg in x.lower()), stdout)):
+		logger.debug(u"Looks like we were already installed in this version. Forcing a reinstall.")
+		force = True
 
-	logger.debug(u"Target: %s, executing pip install %s --ignore-reinstalled --force-reinstall --no-deps" % (target, install_arg))
-	pip_args += ["--ignore-installed", "--force-reinstall", "--no-deps"]
-
-	returncode, stdout, stderr = pip_caller.execute(*pip_args)
-	if returncode != 0:
-		raise exceptions.UpdateError("Error while executing pip install --force-reinstall", (stdout, stderr))
+	if force:
+		logger.debug(u"Target: %s, executing pip install %s --ignore-reinstalled --force-reinstall --no-deps" % (target, install_arg))
+		pip_args += ["--ignore-installed", "--force-reinstall", "--no-deps"]
+	
+		returncode, stdout, stderr = pip_caller.execute(*pip_args)
+		if returncode != 0:
+			raise exceptions.UpdateError("Error while executing pip install --force-reinstall", (stdout, stderr))
 
 	return "ok"
